refactor!: cmd.core -> utils.run
