--- docs/cmd/core.md
@@ -1,8 +1,0 @@
-# `libvcs.cmd.core`
-
-```{eval-rst}
-.. autoapimodule:: libvcs.cmd.core
-   :members:
-   :show-inheritance:
-   :undoc-members:
-```
