+++ docs/cmd/index.md
@@ -19,5 +19,4 @@ versions.
 git
 hg
 svn
-core
 ```
