+++ docs/internals/index.md
@@ -14,5 +14,6 @@ exc
 types
 dataclasses
 query_list
+run
 subprocess
 ```
