+++ docs/internals/run.md
@@ -1,0 +1,9 @@
+# `libvcs.utils.run`
+
+```{eval-rst}
+.. autoapimodule:: libvcs.utils.run
+   :members:
+   :show-inheritance:
+   :undoc-members:
+   :exclude-members: StrOrBytesPath, StrPath, logger
+```
