+++ libvcs/cmd/git.py
@@ -3,7 +3,7 @@ import shlex
 from typing import Any, Literal, Optional, Sequence, Union
 
 from ..types import StrOrBytesPath, StrPath
-from .core import run
+from libvcs.utils.run import run
 
 _CMD = Union[StrOrBytesPath, Sequence[StrOrBytesPath]]
